+++ myBank.h
@@ -7,7 +7,7 @@ void balance(int x);
 void deposit(int x, double y);
 void withdrawal(int x, double y);
 void closeAcc(int x);
-void addInterest(double x);
+void addInterest(int x);
 void printAllAccs();
 void closeAllAccs();
